Move isSingleSliceGA4 to isSingleSlice in GA4 module.
