+++ assets/js/modules/analytics-4/utils/chart.js
@@ -116,3 +116,33 @@ export function extractAnalyticsDataForPieChart( report, options = {} ) {
 
 	return dataMap;
 }
+
+/**
+ * Checks if there is a single row of data or one row of the provided GA4 report is contributing 100% of the total for a given dimension.
+ *
+ * Note that chart reports will be in the multi-date range format.
+ *
+ * @since n.e.x.t
+ *
+ * @param {Object} report The report data object.
+ * @return {(boolean|undefined)} Returns undefined if report is undefined, true/false for the above conditions.
+ */
+export const isSingleSlice = ( report ) => {
+	if ( report === undefined ) {
+		return undefined;
+	}
+
+	const currentDateRangeRows = ( report?.rows || [] ).filter(
+		( { dimensionValues } ) => dimensionValues[ 1 ].value === 'date_range_0'
+	);
+
+	if (
+		currentDateRangeRows?.length === 1 ||
+		currentDateRangeRows?.[ 0 ]?.metricValues?.[ 0 ]?.value ===
+			report?.totals?.[ 0 ]?.metricValues?.[ 0 ]?.value
+	) {
+		return true;
+	}
+
+	return false;
+};
